move is_sparse() from file.cpp to mmap.cpp
