+++ include/libtorrent/file.hpp
@@ -88,10 +88,6 @@ namespace libtorrent {
 	using handle_type = int;
 #endif
 
-#ifdef TORRENT_WINDOWS
-	bool is_sparse(HANDLE file);
-#endif
-
 	struct TORRENT_EXTRA_EXPORT directory
 	{
 		directory(std::string const& path, error_code& ec);
